Create a new file to deal with Raft WAL.
